Allow preview when node is directory
Currently if you enable preview on a file, it works as expected. As you move up/down, it previews, the files, and on a directory it shows nothing. The directory preview _works_, just shows nothing

However, toggling preview when a directory is selected will simply do nothing, which means even when you then move to a file, preview doesn't work. This is frustrating for event handlers and commands which enable preview (see #783)